+++ src/plugins/core/backgrounds/ImageSettings.tsx
@@ -7,8 +7,13 @@ interface Props {
   onChange: (settings: Settings) => void;
 }
 
-class ImageSettings extends React.PureComponent<Props> {
+interface State {
+  oversized?: boolean;
+}
+
+class ImageSettings extends React.PureComponent<Props, State> {
   static defaultProps = Image.defaultProps;
+  state: State = {};
 
   render() {
     return (
@@ -21,12 +26,17 @@ class ImageSettings extends React.PureComponent<Props> {
           />
         </label>
 
+        {this.state.oversized && <p className="info">Large images may affect performance</p>}
+
         <img src={this.props.image} style={{width: '100%', height: 'auto'}} />
       </div>
     );
   }
 
   private loadImage(files: FileList) {
+    // Check for oversized
+    this.setState({ oversized: files[0].size > 2097152 }); // Larger than 2mb
+
     // Setup file reader
     const reader = new FileReader();